107 create user detail for admin (#117)
* changes laundry selection when changing org

* add function to add admin as manager of org

* remove reservations of user when expelling

* fix security breach : user can now only delete their reservations

* implement display of reservation details for admins

+++ calm-webserver/app/Http/Controllers/ReservationController.php
@@ -231,7 +231,14 @@ class ReservationController extends Controller
      */
     public function destroy(string $id)
     {
-        $res = Reservation::find($id);
+        $res = Reservation::findOrFail($id);
+
+        if($res->user->id != Auth::user()->id){
+            return back()->withErrors([
+                "La réservation n'existe pas"
+            ]);
+        }
+
         $res->delete();
         return redirect()->route('reservations.index');
     }
